Pi/requirements.txt
aws==0.2.5
awscrt==0.13.9
awsiotsdk==1.11.0
bcrypt==3.2.0
boto==2.49.0
cffi==1.15.0
cryptography==37.0.0
fabric==2.7.0
invoke==1.7.0
paramiko==2.10.4
pathlib2==2.3.7.post1
pip==22.0.4
prettytable==3.2.0
pycparser==2.21
PyNaCl==1.5.0
pyserial==3.5
setuptools==62.1.0
six==1.16.0
wcwidth==0.2.5
wheel==0.37.1
